Unify window dims to uint32_t

+++ src/WSI/Glfw.h
@@ -26,7 +26,7 @@ int messageLoop( PlatformWindow window );
 
 bool platformPresentationSupport( VkInstance instance, VkPhysicalDevice device, uint32_t queueFamilyIndex, PlatformWindow window );
 
-PlatformWindow initWindow( const std::string& name, int canvasWidth, int canvasHeight );
+PlatformWindow initWindow( const std::string& name, uint32_t canvasWidth, uint32_t canvasHeight );
 void killWindow( PlatformWindow window );
 
 VkSurfaceKHR initSurface( VkInstance instance, PlatformWindow window );
@@ -217,12 +217,12 @@ bool platformPresentationSupport( VkInstance instance, VkPhysicalDevice device, 
 	return glfwGetPhysicalDevicePresentationSupport( instance, device, queueFamilyIndex ) == GLFW_TRUE;
 }
 
-PlatformWindow initWindow( const std::string& name, const int canvasWidth, const int canvasHeight ){
+PlatformWindow initWindow( const std::string& name, const uint32_t canvasWidth, const uint32_t canvasHeight ){
 	std::string title = name + " -- GLFW";
 
 	glfwWindowHint( GLFW_CLIENT_API, GLFW_NO_API );
 	glfwWindowHint( GLFW_VISIBLE, GLFW_FALSE );
-	GLFWwindow* window = glfwCreateWindow( canvasWidth, canvasHeight, title.c_str(), NULL, NULL );
+	GLFWwindow* window = glfwCreateWindow( static_cast<int>(canvasWidth), static_cast<int>(canvasHeight), title.c_str(), NULL, NULL );
 
 	if( !window ) throw "Trouble creating GLFW window!";
